+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Add UnboundStack (like Stack, but allows hit-testing (e.g. clicking, hovering, ...) Widgets outside of its bounds)
 * Add RenderUnboundStack (the RenderObject which enabled all the features of UnboundStack)
+* Add AdaptiveRefreshIndicator.slivers (allows to use multiple slivers)
 * Add Example project to improve the overall documentation
 
 ### Changed ###